feat(breadcrumbs): component for note parents navigation
